+++ tests/conftest.py
@@ -77,3 +77,15 @@ def reset_logger_context() -> None:
     import tomodachi.logging
 
     tomodachi.logging.reset_context()
+
+
+@pytest.fixture(scope="function", autouse=True)
+def uninstrument_opentelemetry() -> None:
+    from tomodachi.opentelemetry import TomodachiInstrumentor
+    from tomodachi.opentelemetry.distro import OpenTelemetryConfigurator, OpenTelemetryDistro
+
+    OpenTelemetryConfigurator().reset()
+    OpenTelemetryDistro().reset()
+
+    instrumentor = TomodachiInstrumentor()
+    instrumentor.uninstrument()