+++ tomodachi/opentelemetry/__init__.py
@@ -1,0 +1,5 @@
+from tomodachi.opentelemetry.instrumentation import TomodachiInstrumentor
+
+__all__ = [
+    "TomodachiInstrumentor",
+]